DBprotection: updateShield 99.4->100 (the (s32) on the rotZ u16-store drops while shieldAngle's (u16)(s32) stays - per-site cast ownership read off T's clrlwi positions)

+++ src/main/dll/DB/DBprotection.c
@@ -946,7 +946,7 @@ void DBprotection_updateShield(int *obj)
     state->shieldSfxLatch = 0;
   }
 
-  *(u16 *)&((GameObject *)obj)->anim.rotZ = (s32)(lbl_803E57DC * angleCos);
+  *(u16 *)&((GameObject *)obj)->anim.rotZ = lbl_803E57DC * angleCos;
   state->shieldAngle = (u16)(s32)(lbl_803E57E0 * timeDelta + (f32)state->shieldAngle);
 }
